requirements.txt
asgiref==3.5.2
Django==4.0.5
django-filter==21.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
gunicorn==20.1.0
Markdown==3.3.7
pipfile==0.0.2
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
whitenoise==6.1.0